requirements.txt
pandas
pyarrow
shapely
openpyxl
google-cloud-bigquery
confluent-kafka
kafka-python
pyspark
